+++ app/chrome-extension/entrypoints/background/tools/browser/index.ts
@@ -22,6 +22,8 @@ export { javascriptTool } from './javascript';
 export { pasteTextTool } from './paste-text';
 export { consoleTool } from './console';
 export { fileUploadTool } from './file-upload';
+export { pasteImageTool } from './paste-image';
+export { formValueTool } from './form-value';
 export { readPageTool } from './read-page';
 export { computerTool } from './computer';
 export { postToXTool } from './post-to-x';